Move confirm processing to concern module
